LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/writing-mode-parsing-sideways-rl-001-expected.txt


FAIL writing-mode: sideways-rl assert_equals: expected "sideways-rl" but got "horizontal-tb"
